Allocated the flywheels to a manual control.
